requirement.txt
numpy
matplotlib
pygame
groq
langchain
langchain-groq
langgraph
pyyaml
python-dotenv
piper-tts
sounddevice
scipy
pvporcupine
langchain_community
ddgs
webrtcvad
setuptools
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
chromadb
sentence-transformers
neo4j

langgraph-checkpoint-sqlite
todoist_api_python
qdrant_client
langchain_google_genai
langchain
langchain-qdrant
google-genai
langchain-google-genai
pgvector
psycopg2-binary